refactor(tui): extract event reducer into state/events.rs
Move the event half of the reducer — reduce_inner (the AppEvent dispatch),
apply_update (folds a session update into a pane), and the one_line notice
helper — into state/events.rs. reduce_inner is pub(super); mod.rs's public
reduce() wrapper calls it. mark_shown_seen stays at the root (shared with
keys). state/mod.rs is now just AppState + its accessors, the reduce()
wrapper, shared consts/helpers, and the test module. No behavior change.
